+++ ft-sdk/CHANGELOG.md
@@ -90,6 +90,14 @@
 1. Support for enabling session replay recording feature
 
 ---
+# agent 1.6.15-alpha03
+1. WebView data collection supports Tencent X5
+
+---
+# agent 1.6.15-alpha02
+1. Added background field to error_situation
+
+---
 # agent 1.6.15-alpha01
 1. Fixed the incorrect way of retrieving the cursor count.